Create Example4.py
Using an array to extend the scope of Fizz Buzz - it now handles Fizz, Buzz, Quack, Moo, Splat and Kapowee using less code than before and one array of noises to make. Note that this is an array of tuples - where the tuple holds the number to divide by and the noise to display.